pyproject.toml
[project]
name = "arabia"
version = "0.1.0"
description = "A FastAPI backend for the Arabia project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv==1.0.1",
    "fastapi==0.115.11",
    "mistral>=19.0.0",
    "pinecone==6.0.2",
    "requests>=2.32.3",
    "uvicorn==0.34.0",
    "pydantic-settings==2.8.1",
    "appwrite==9.0.2",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "certifi==2025.1.31",
    "click==8.1.8",
    "eval-type-backport==0.2.2",
    "exceptiongroup==1.2.2",
    "h11==0.14.0",
    "httpcore==1.0.7",
    "httpx==0.28.1",
    "idna==3.10",
    "jsonpath-python==1.0.6",
    "mypy-extensions==1.0.0",
    "pinecone-plugin-interface==0.0.7",
    "pydantic[email]==2.10.6",
    "pydantic-core==2.27.2",
    "python-dateutil==2.9.0.post0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "starlette==0.46.1",
    "typing-inspect==0.9.0",
    "typing-extensions==4.12.2",
    "urllib3==2.3.0",
    "gunicorn==21.2.0",
    "beautifulsoup4>=4.12.2",
    "mistralai>=1.6.0",
    "pytest-mock>=3.14.0",
]
